+++ src/ast/markdown-to-ast.ts
@@ -1,18 +1,33 @@
-import { Nodes } from "npm:@types/mdast";
+import { Node, Nodes } from "npm:@types/mdast";
 import { fromMarkdown } from "npm:mdast-util-from-markdown";
 import { gfmFromMarkdown } from "npm:mdast-util-gfm";
 import { gfm } from "npm:micromark-extension-gfm";
 import * as R from "npm:ramda";
+import { readAllFromStdin } from "../io/read-all-from-stdin.ts";
+
+/**
+ * Converts the given Markdown to an AST, without any `position` properties.
+ * @param markdown The Markdown to convert to an AST.
+ * @returns The given Markdown converted to an AST.
+ */
+export function markdownToAst(markdown: string): Nodes {
+  const ast = fromMarkdown(markdown, {
+    extensions: [gfm()],
+    mdastExtensions: [gfmFromMarkdown()],
+  });
+  return removePosition(ast);
+}
 
 /**
  * Clone an object, excluding "position" properties
  */
-const cloneWithoutPosition = R.compose(R.omit(["position"]), R.clone);
+const cloneWithoutPosition: <T extends Node>(ast: T) => Omit<T, "position"> = R
+  .compose(R.omit(["position"]), R.clone);
 
 /**
  * Clone an AST, excluding "position" properties
  */
-export function removePosition(node: unknown) {
+export function removePosition<T extends Node>(node: T): Omit<T, "position"> {
   if (R.is(Array, node)) {
     return R.map(removePosition, node);
   }
@@ -22,17 +37,9 @@ export function removePosition(node: unknown) {
   return node;
 }
 
-export function markdownToAst(markdown: string): Nodes {
-  const ast = fromMarkdown(markdown, {
-    extensions: [gfm()],
-    mdastExtensions: [gfmFromMarkdown()],
-  });
-  return removePosition(ast);
-}
-
 // read from Deno.stdin, console.log the result
 if (import.meta.main) {
-  const inputMarkdown = await Deno.readTextFile("/dev/stdin");
+  const inputMarkdown = await readAllFromStdin();
   const ast = markdownToAst(inputMarkdown);
   const output = JSON.stringify(ast, null, 2);
   console.log(output);
