+++ src/markdown/format-code.ts
@@ -2,6 +2,12 @@ import { run } from "run_simple";
 import { global, groups, sequence } from "../strings/regex.ts";
 import { BOX_REGEX } from "../model/box.ts";
 
+/**
+ * Format source code using Deno's built-in formatter.
+ * @param ext the file extension of the source code
+ * @param code the source code to format
+ * @returns the formatted source code
+ */
 export async function formatCode(
   ext: "md" | "json" | "jsonc" | "ipynb" | "ts" | "tsx" | "js" | "jsx",
   code: string,
@@ -11,7 +17,7 @@ export async function formatCode(
   })).trimEnd() + "\n";
 }
 
-function removeBackslashBeforeBox(code: string): string {
+export function removeBackslashBeforeBox(code: string): string {
   return code.replaceAll(
     global(sequence(/(?<hashOrDash>[#-] )/, "\\", BOX_REGEX)),
     (...args) =>
